requirements_face_recognition.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
tensorflow==2.15.0

